requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML and Computer Vision
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Face Detection
insightface==0.7.3
onnxruntime==1.16.3

# License Plate Detection
ultralytics==8.0.220

# Frontend
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

